backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
bcrypt==4.1.3
python-multipart==0.0.20
pandas==2.3.2
openpyxl==3.1.5
httpx==0.28.1
reportlab==4.4.3
jinja2==3.1.6
